int para double

+++ Source Code Inspection/src/br/calebe/ticketmachine/core/TicketMachine.java
@@ -10,16 +10,16 @@ import java.util.Iterator;
  */
 public class TicketMachine {
 
-    protected int valor;
-    protected int saldo;
+    protected double valor;
+    protected double saldo;
     protected int[] papelMoeda = {2, 5, 10, 20, 50, 100};
 
-    public TicketMachine(int valor) {
+    public TicketMachine(double valor) {
         this.valor = valor;
         this.saldo = 0;
     }
 
-    public void inserir(int quantia) throws PapelMoedaInvalidaException {
+    public void inserir(double quantia) throws PapelMoedaInvalidaException {
         boolean achou = false;
         for (int i = 0; i < papelMoeda.length && !achou; i++) {
             if (papelMoeda[i] == quantia) {
@@ -32,11 +32,11 @@ public class TicketMachine {
         this.saldo += quantia;
     }
 
-    public int getSaldo() {
+    public double getSaldo() {
         return saldo;
     }
 
-    public int getTroco() {
+    public double getTroco() {
         return (saldo-valor);
     }
